Apply workaround for the >32 cores per port
Issue #1072 have a proposed workaround to cast 1 to uint64_t to properly support more than 32 cores per port.

Apply that fix to this fork, however it is probably better to rewrite that part of the code to be more scalable

+++ src/stx/stl/trex_stl_streams_compiler.h
@@ -83,7 +83,7 @@ public:
             return false;
         }
         /* highest bit pushed to left and then -1 will give all the other bits on */
-        return ( (dp_core_mask & ( (1 << dp_core_count) - 1 ) ) != 0);
+        return ( (dp_core_mask & ( ( (uint64_t)1 << dp_core_count) - 1 ) ) != 0);
     }
 
 private:
